switches back to storing uint8 observations. Some kind of memory bug when saving replay. Disabling for now

+++ dqn/memory/memory.py
@@ -18,10 +18,11 @@ class ExperienceBuffer:
         self.next_idx = 0
         self.nb_samples = 0
 
-        self.observations   = torch.empty(self.capacity, *(obs_shape), dtype=torch.float) # uint8
+        self.observations   = torch.empty(self.capacity, *(obs_shape), dtype=torch.uint8) # uint8
         self.actions        = torch.empty(self.capacity, dtype=torch.int64)
         self.rewards        = torch.empty(self.capacity, dtype=torch.float)
         self.dones          = torch.empty(self.capacity, dtype=torch.bool)
+
 
         self.pending_effect = False
